Move "actions.rs" to "matcher.rs" and document.
